* Applications, configurable pipelines, bug squishing
* The examine_dist and find_modules methods now do their
work by looking for the tasks they should do in
examine_dist_techniques and find_modules_techniques.
This makes it easy to subclass and change the indexing,
as in MyCPAN::Indexer::DPAN.

* The applications in *::App::* can now set their own
defaults. This way you can get around making config
files and using command line arguments.

* Added two new techniques to find module files:
look in META.yml 'provides', or search all of the
subdirs for *.pm. This was important for Term::Readline::Perl

* Fixed various impedence issues where modules need
to join each other.