* Queue just looks in backpan_dir, DPAN.pm get docs
The Queue class now just looks in the backpan_dir
configuration. It's up to the application to the
application to populate that through configuration
or defaults. If It's not set, App::BackPAN::Indexer
sets it to @ARGV

I added the docs to the DPAN app. I just need to
make it work now.