* App::DPAN works, but there's a little polish to do
Most of the the MyCPAN::App::DPAN works, along with the
dpan.pl script that uses it. There are still a few indexing
oddities, such as pmtools, but other than that it's time
to test it on Windows.

I changed the App::BackPAN::Indexer to handle the command
line a bit differently. If backpan_dir is not in the config,
it looks at @ARGV. If there is nothing in @ARGV, it uses
the current working directory.

Now, backpan_dir might be a single element or an array
reference. It should always be one thing, but I'm changing
it in the middle here so I have to support the single
element (i.e. not array ref) until I find all the other
places that use it. There is some ugly code to deal with
it being possibly both.