* Get rid of Tk prereq, fix up writing to modules/
The Tk stuff is not a prereq, but you need to install it to
use the Tk interface.

In DPAN.pm, I needed to ensure that the modules/ directory
existed before I tried to write to it, and that everyone
was using the right file paths. This came up when I decided
to index any directories, just not those that already
looked like a CPAN directory.

For some reason, the exporting from File::Find::Closures
isn't working under 5.10, so I'll just use the full
package specification and worry about it later.