* make paths absolute, put temp dirs in the current working dir
I ran into a lot of path problems on Windows that didn't show
up elsewhere. I eventually solved these by running the paths
through rel2abs. Also, I create the temporary directories
in the current working directory.

I also needed to fix up the ARGV processing. The command-line
arguments become arguments to run, but getopts needs them to
be in @ARGV. So, for a quick fix, localize @ARGV, then after
getopts changeds it, assign the result back to the original
list. Use that as the backpan_dirs if that is not already
set in the config.