* Archive::Extract reports failure when it works, so ignore it
For some reason, Archive::Extract 0.26 on Windows under cygwin
and Starwberry Perl doesn't think it works. extract() returns
false although it works. Instead of checking the return value,
I'll record it and move on as if it worked.

I haven't been able to reproduce the problem though. Maybe it
doesn't have anything to do with tempdirs.