* Get rid of "meta" as a literal, clean up temp directories
I don't like that "meta" was a special string all over the
place. That name of the directory for reports is now
an undocumented configuration directive.

I added some bits to do cleanup on the way out, including
removing all of the temporary directories. I wish I could
let File::Temp do that, but in odd cases it seems to clean
up some directories too early. Maybe the forked processes
are doing that.