* in dpan, only process the yml files, and only if the dist is still there
Use the new CPAN::PackageDetails, which sorts the package names. The
new App::DPAN puts relative paths in there is it sees the author/id
structure.

I've continued on the work to decide what should show up in 02packages.
There's a catch to this so far: you have to be in the same directory
structure as the original indexing for this to work because it's looking
for the same files. The index results have the full path the dist that
it indexed. If you try it with another directory structure, even the
tiniest bit different in any part of the path, you'll still pick
up the wrong paths from the original index.

That's just a temporary problem though and I haven't decided how
to fix it. For instance, I can just look at the filename, ensure
the dist has the same checksums, then pull out the right info. I
haven't quite decided how I like that.