* Check $Notes->{Finished} to see if the interface shuold keep going.
Previously I was checking $Notes->{Left} to see if the interface
should keep going. However, that was computed in the dispatcher
parent process when it launched a child. Consequently, it went
down to zero when it launched the last process even though that
process had not completed. The child process would complete, but
the rest of dpan went on to create the 02packages. That happens
fast enought that the final report doesn't make it into the
processing. Oops.