* Hard-code some package names to skip automatically, don't complain about Archive::Extract on Windows
We want to skip some packages. There really isn't a good way to do this
because there is no rule that encapsulates it. You juast have to
know what not to put in 02pacakges. For instance, don't ever index
a package called "main".

Archive::Extract seems to be working just fine in Windows, but it still gives
and error. I'll ignore that error if it is on Windows.