* Work out get_config conflict in Coordinator and Indexer
The get_config in Indexer should realyl be init_config and only
called once to install the defaults.

There's a secondary problem about storing a list of dirs
for backpan_dir. ConfigReader::Simple only sets values that
are strings. We'll figure out how to get around that later. For
now we just have to know it's a space separated string and hard-code
around that.