Remove fixed path.
Use File::Spec and File::HomeDir to point to the
user home directory for the tied hash database.
Since File::HomeDir->my_desktop could return undef if the desktop is not
available, the method defaults to the user's home directory if the
desktop is not found.

Added dependency to Makefile.PL, using version 0.07 as lower since it is
the one with the older API interface.