Refactor if-else branch.
The eval exception ($at) is already included in the error log, and all
the cases were returning (stopping) setting different run_info depending
on the fatality-ness of the exception. Therefore the triple if-elsif
branch can be compressed in a single branch with the discrimination on
the run_info information and a single log/error statement that informs
about the "stopping" of the execution.